+++ products/AbstractBuilder.node.js
@@ -120,17 +120,10 @@ class AbstractBuilder extends jtree.TreeNode {
       //Assert
       equal(errs.length, 0, "should be no errors")
     }
-    const checkGrammarExamples = (equal, program) => {
-      // Act
-      const exampleErrors = program.getErrorsInGrammarExamples()
-      if (exampleErrors.length) console.log(exampleErrors)
-      // Assert
-      equal(exampleErrors.length, 0, exampleErrors.length ? "examples errs: " + exampleErrors : "no example errors")
-    }
     const handGrammarProgram = new jtree.HandGrammarProgram(Disk.read(grammarPath))
     testTree[`grammarCheckOf${grammarPath}`] = equal => checkGrammarFile(equal, jtree.compileGrammarAndCreateProgram(grammarPath, __dirname + "/../langs/grammar/grammar.grammar"))
     testTree[`handGrammarCheckOf${grammarPath}`] = equal => checkGrammarFile(equal, handGrammarProgram)
-    testTree[`handGrammarExamplesCheckOf${grammarPath}`] = equal => checkGrammarExamples(equal, handGrammarProgram)
+    Object.assign(testTree, handGrammarProgram.examplesToTestBlocks())
     return testTree
   }
   _help(filePath = process.argv[1]) {